+++ tests/src/Opulence/Events/Mocks/Event.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2016 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+namespace Opulence\Tests\Events\Mocks;
+
+/**
+ * Defines a mock event
+ */
+class Event
+{
+    // Don't do anything
+}