+++ tests/src/Opulence/Events/Mocks/Listener.php
@@ -9,7 +9,6 @@
 namespace Opulence\Tests\Events\Mocks;
 
 use Opulence\Events\Dispatchers\IEventDispatcher;
-use Opulence\Events\IEvent;
 
 /**
  * Mocks an event listener
@@ -19,11 +18,11 @@ class Listener
     /**
      * Mocks a listen function that does nothing
      *
-     * @param IEvent $event The event to handle
+     * @param object $event The event to handle
      * @param string $eventName The name of the event
      * @param IEventDispatcher $dispatcher The event dispatcher
      */
-    public function doNothing1(IEvent $event, $eventName, IEventDispatcher $dispatcher)
+    public function doNothing1($event, $eventName, IEventDispatcher $dispatcher)
     {
         // Don't do anything
     }
@@ -31,24 +30,12 @@ class Listener
     /**
      * Mocks a listen function that does nothing
      *
-     * @param IEvent $event The event to handle
+     * @param object $event The event to handle
      * @param string $eventName The name of the event
      * @param IEventDispatcher $dispatcher The event dispatcher
      */
-    public function doNothing2(IEvent $event, $eventName, IEventDispatcher $dispatcher)
+    public function doNothing2($event, $eventName, IEventDispatcher $dispatcher)
     {
         // Don't do anything
-    }
-
-    /**
-     * Mocks a listen function that stops propagation
-     *
-     * @param IEvent $event The event to handle
-     * @param string $eventName The name of the event
-     * @param IEventDispatcher $dispatcher The event dispatcher
-     */
-    public function stopsPropagation(IEvent $event, $eventName, IEventDispatcher $dispatcher)
-    {
-        $event->stopPropagation();
     }
 }